refactor: introduce ObjectsTrait
